move RegisterForm to components/
