+++ pages/strom/register.tsx
@@ -1,7 +1,7 @@
 import {NextPage} from 'next'
 
 import {PageLayout} from '../../src/components/PageLayout/PageLayout'
-import {RegisterForm} from '../../src/pages/RegisterForm/RegisterForm'
+import {RegisterForm} from '../../src/components/RegisterForm/RegisterForm'
 
 const Register: NextPage = () => (
   <PageLayout seminarId={1}>
